fix: tips component
